+++ app/src/main/java/com/haarman/pebblenotifier/notifications/strategies/DefaultNotificationTextStrategy.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.haarman.pebblenotifier.notifications;
+package com.haarman.pebblenotifier.notifications.strategies;
 
 import android.app.Notification;
 import android.content.Context;
